Restructure api files and handle query params
When moving the makeUrl function, it was actually updated to handle
optional query parameters. That should have been a commit on its own,
but it has been included here instead.
