--- src/api/api.js
@@ -1,12 +1,0 @@
-// This is a temporary testing URL
-const baseUrl = 'http://forum.studiogoose.dk';
-
-export const alwaysHeaders = {
-  'Content-Type': 'application/json',
-  Accept: 'application/json',
-  'X-Requested-With': 'XMLHttpRequest',
-};
-
-export function makeUrl(route) {
-  return `${baseUrl}${route}`;
-}
