+++ src/api/constants.js
@@ -1,0 +1,8 @@
+// This is a temporary testing URL
+export const baseUrl = 'http://forum.studiogoose.dk';
+
+export const alwaysHeaders = {
+  'Content-Type': 'application/json',
+  Accept: 'application/json',
+  'X-Requested-With': 'XMLHttpRequest',
+};
